# Device optimization for axom::Array