# Summary

Currently, `axom::Array<T>` initializes memory on the device by creating an array on the host, then copying it to the device. This is to work around limitations where `T` has a non-trivial default constructor, as that constructor would need to be device-annotated.

However, if `T` is trivially default-constructible, the host-allocate-and-copy workaround is unnecessary and may cost some performance. In this PR, we launch a kernel for default-initialization if the type `T` is trivially default-constructible.

As a bonus, we replace the use of SFINAE in `detail::ArrayOpsBase` with tag-dispatch.

## Performance

Measured on rzansel with `./bin/mesh_tester --method [accel_method] -p 3 --skipWeld -i [test_mesh]`:

| Test case | `develop` | this PR | speedup |
| ---------- | ---------- | ------- | -------- |
| car.stl - bvh | 0.358s | 0.248s | 1.45x |
| car.stl - implicit | 2.439s | 1.413s | 1.73x | 
| car.stl - uniform | 2.643s | 0.795s | 3.32x |
| porsche.stl - bvh | 0.131s | 0.128s | 1.03x |
| porsche.stl - implicit | 0.207s | 0.159s | 1.3x |
| porsche.stl - uniform | 0.277s | 0.172s | 1.61x |